services/session-orchestrator/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
modal-client==0.62.181
boto3==1.34.97
httpx==0.27.0
